requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
contourpy==1.2.0
cycler==0.12.1
Django==5.0.3
djangorestframework==3.15.1
flatbuffers==24.3.25
fonttools==4.50.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10
imbalanced-learn==0.12.0
imblearn==0.0
joblib==1.3.2
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.26.4
opt-einsum==3.3.0
optree==0.11.0
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
pyparsing==3.1.2
PyQt5==5.15.10
rich==13.7.1
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
setuptools==68.2.2
sqlparse==0.4.4
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==2.4.0
threadpoolctl==3.4.0
tzdata==2024.1
Werkzeug==3.0.1
wheel==0.41.2
wrapt==1.16.0
